+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed property descriptors of some `Math` and `Number` constants
 - Added detection of NodeJS [bug](https://github.com/nodejs/node/issues/41038) in `structuredClone` that can not clone `DOMException` (just in case for future versions that will fix other issues)
 - Added NodeJS 18.3 compat data mapping
+- Updated and fixed Deno 1.21 and 1.22 compat data mapping
 - Updated Electron 20.0 compat data mapping
 
 ##### [3.22.8 - 2022.06.02](https://github.com/zloirock/core-js/releases/tag/v3.22.8)